projects/sketch_to_image_diffusion/requirements.txt
# Sketch-to-Image Diffusion Generator
# CPU-optimized implementation

# Core ML
torch>=2.0.0
torchvision>=0.15.0

# UI
gradio>=4.0.0

# Utilities
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
matplotlib>=3.7.0

# Optional: for faster CPU inference
# torch-compile support built into PyTorch 2.0+
